@@ -8,7 +8,6 @@ import android.content.Intent;
 import android.graphics.Bitmap;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.text.method.HideReturnsTransformationMethod;
 import android.text.method.PasswordTransformationMethod;
 import android.view.Gravity;
@@ -21,6 +20,8 @@ import com.petkit.android.utils.CommonUtils;
 
 import java.util.Iterator;
 import java.util.Map;
+
+import androidx.fragment.app.Fragment;
 
 public class BaseUIUtils<T> {
 
