@@ -1,10 +1,11 @@
 package com.petkit.matetool.widget;
 
 import android.content.Context;
-import android.support.v4.view.ViewPager;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 import android.view.ViewParent;
+
+import androidx.viewpager.widget.ViewPager;
 
 
 /**
