@@ -1,12 +1,14 @@
 package com.vilyever.socketclient.server;
 
-import android.support.annotation.NonNull;
 
 import com.vilyever.socketclient.SocketClient;
 import com.vilyever.socketclient.helper.SocketClientAddress;
 import com.vilyever.socketclient.helper.SocketConfigure;
 
 import java.net.Socket;
+
+import androidx.annotation.NonNull;
+
 
 /**
  * SocketServerClient
